requirement.txt
asgiref==3.2.7
Django==3.0.3
djangorestframework==3.10.3
flake8==3.6.0
mccabe==0.6.1
pillow>=5.3.0,<5.4.0
pycodestyle==2.4.0
pyflakes==2.0.0
pytz==2019.3
sqlparse==0.3.1
